more updates/fixes/renames
