+++ packages/desktop-client/src/style/colors.ts
