+++ packages/desktop-client/src/style/themes/light.ts
@@ -1,4 +1,4 @@
-import * as colorPalette from '../colors-new';
+import * as colorPalette from '../palette';
 
 export const pageBackground = colorPalette.gray100;
 export const pageBackgroundModalActive = colorPalette.gray200;
